# perf(linter/jsx-no-comment-textnodes): remove regex for checking comment patterns