@camchenry 2024-10-14
@DonIsaac i think this is ready to be review. I was going to wait until we had the small file benchmark, but I think there's enough data to show this is most likely a perf win.